requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-socketio==5.11.1

# Alpaca SDK
alpaca-py==0.28.3

# Database
motor==3.6.0
pymongo==4.10.1

# Redis (optional)
redis==5.2.0

# Configuration
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Production
gunicorn==23.0.0

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
